pyproject.toml
[project]
name = "depth-anything-v2"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "albumentationsx>=2.0.10",
    "decord>=0.6.0",
    "easydict>=1.13",
    "einops>=0.8.1",
    "h5py>=3.14.0",
    "imageio[ffmpeg]>=2.37.0",
    "ipython>=9.4.0",
    "matplotlib>=3.10.1",
    "numpy>=2.1.2",
    "opencv-python>=4.11.0.86",
    "pillow>=11.2.1",
    "torch>=2.4.0",
    "torchvision>=0.20.0",
    "tqdm>=4.66.5",
    "transformers>=4.57.1",
    "wandb>=0.19.9",
    "xformers>=0.0.29.post1",
]

[[tool.uv.index]]
name = "pytorch-cu121"
url = "https://download.pytorch.org/whl/cu121"
explicit = true

[tool.uv.sources]
torch = [
  { index = "pytorch-cu121", marker = "sys_platform == 'linux' or sys_platform == 'win32'" },
]
torchvision = [
  { index = "pytorch-cu121", marker = "sys_platform == 'linux' or sys_platform == 'win32'" },
]

[tool.basedpyright]
strictListInference = true
strictDictionaryInference = true
strictSetInference = true
strictParameterNoneValue = true
deprecateTypingAliases = true
enableExperimentalFeatures = true
strictGenericNarrowing = true
